% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-newbler.R
\name{process.orf.fasta.file}
\alias{process.orf.fasta.file}
\title{Import contig data from a fasta file}
\usage{process.orf.fasta.file(orf.fa,mainSplitString)}
\arguments{
\item{orf.fa}{'orf.fa' is a .fasta file generated by MetaGeneMark, containing the DNA sequence from each predicted ORF}
\item{mainSplitString}{'mainSplitString' (character string) gives you flexibility about how to split up the ORF header...seems to be some variability in how MetaGeneMark encodes this?}
}
\value{
A data.frame containing three columns (contig /cov/erage, contig /len/ength and cognate ORF-id (simplified)). Row names are ORF identifiers.
}
\description{
Extracts basic information about ORFs obtained by MetaGeneMark
}
\examples{
\dontrun{a<-extract.contig.data()}
\dontrun{b<-process.orf.fasta.file("UM-32-0_nucl_mod.fasta")}
}

