% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-newbler.R
\name{read.tax}
\alias{read.tax}
\title{Read in a .tax output file from LCAMapper}
\usage{
read.tax(file)
}
\arguments{
\item{file}{is a file with name *-tax.file output from lcamapper.sh}
}
\value{
A character vector containing ORF data and post-LCA annotations. Each element contains results from one ORF.
}
\description{
Import taxonomic assignments from an ORF-level homology search analysis processed by lcamapper.sh
}
\examples{
\dontrun{a<-read.tax("UM-32-0_ORF-tax.txt")}
}

